requirements.txt
# pip install requirements.txt
pandas
matplotlib
networkx
numpy==1.16
plotly==4.5.0
Pillow
psutil
opencv-python
imageio
tqdm
scipy
scikit-learn
scikit-image==0.14.2
torchfile
kaggle
notebook>=5.3
ipywidgets>=7.2
easydict
bcolz
mxnetcd